Senga2 ops mpi (#1)
* updated loop ranges to use global sizes, making aligned with OPS requirement

* moved allocation of arrays from static to dynamic which are referenced by OPS DATS

* converted tempin to OPS. replaced data type kind parameter to able to write data to hdf5

* added sample file showing stages of ops loop implementation

* pushing cuda and sequential kernel file generated for dfbydx interior kernel for presentation purpose

* corrected the access specifier information for all par loops defined so far

* added code to do first touch for OPS_DATS

* removed array allocation for most of the OPS datasets by SENGA to let OPS handle allocation

* corrected some par loop ranges

* replaced sizes with global variables

* corrected calculation of crin variable in flamin.F90

* corrected range parameter for right side boundary element kernels

* checked MPI version of senga, correct output for drhs,urhs,erhs for 40 ranks case

* added call to print statistics

* test commit

* combined interior,LH,RH and scaling loops in single kernel

* added kernels for d2fdyz

* written few stencil declarations for d2fdxy kernels

* finished stencil declarations for d2fdxy kernels

* completed stencil declarations for d2fdxz kernels

* finished stencil declarations for d2fdyz kernels

* added par loop declarations for kernels written so far for d2fdxy, d2fdxz and d2fdyz

* implemented remaining loops from d2fdxy into OPS

* implemented remaining loops from d2fdxz into OPS

* implemented remaining loops from d2fdyz into OPS, corrected syntax errors for d2fdxy and d2fdxz

* updated files to handle 3D TGV case

* corrected few 3D kernels related functions

* updated range calculations for differentials kernels

* checked the squential OPS version for 3D TGV datasets. Output matching.

* OPS MPI version correctly producing result, matching with complete F90 version of SENGA+

* deleted check which no longer required

* added change to enable SOA, output matched with current branch of fix/fortran_gbl_mdim

* Update senga2.F90

disabled SOA, not giving improvement

* changed print statement to able to build CUDA version

* correcting dim parameter for ops_arg_gbl

* added soft-link for tiled executables

---------

+++ src/ardums.F90
@@ -25,7 +25,7 @@ SUBROUTINE ardums(array,nxphys,nyphys,nzphys,nspec,ispec)
 !   ARGUMENTS
 !   =========
 
-    real(kind=dp), intent(IN OUT)       :: array(nxphys,nyphys,nzphys,nspec)
+    real(kind=8), intent(IN OUT)       :: array(nxphys,nyphys,nzphys,nspec)
     integer, intent(IN OUT)             :: nxphys
     integer, intent(IN)                 :: nyphys
     integer, intent(IN)                 :: nzphys
