+++ src/chkary.F90
@@ -28,7 +28,7 @@ SUBROUTINE chkary(array,  &
 !   ARGUMENTS
 !   =========
 
-    real(kind=dp), intent(IN)       :: array(nchkxl:nchkxr,nchkyl:nchkyr,nchkzl:nchkzr,ncpcmx)
+    real(kind=8), intent(IN)       :: array(nchkxl:nchkxr,nchkyl:nchkyr,nchkzl:nchkzr,ncpcmx)
     integer, intent(IN OUT)         :: nchkxl
     integer, intent(IN OUT)         :: nchkxr
     integer, intent(IN OUT)         :: nchkyl
@@ -46,7 +46,7 @@ SUBROUTINE chkary(array,  &
 
 !   LOCAL DATA
 !   ==========
-    real(kind=dp) :: arrmin,arrmax
+    real(kind=8) :: arrmin,arrmax
     integer :: ic,jc,kc
     integer :: icmin,jcmin,kcmin
     integer :: icmax,jcmax,kcmax
